+++ lib/features/aicycle_buy_me/domain/usecase/get_duplicate_folder_usecase.dart
@@ -4,9 +4,9 @@ import '../../../../network/api_error.dart';
 import '../../data/model/claim_folder_model.dart';
 import '../repository/aicycle_buy_me_repository.dart';
 
-class GetDuplicateFolderUsecase {
+class BuyMeGetDuplicateFolderUsecase {
   final AiCycleBuyMeRepository repository;
-  GetDuplicateFolderUsecase(this.repository);
+  BuyMeGetDuplicateFolderUsecase(this.repository);
   Future<Either<APIErrors, ClaimFolderModel>> call({
     required String externalClaimId,
   }) async {