+++ lib/features/folder_details/domain/usecase/detele_image_by_id_usecase.dart
@@ -3,10 +3,10 @@ import 'package:dartz/dartz.dart';
 import '../../../../network/api_error.dart';
 import '../repository/folder_detail_repository.dart';
 
-class DeleteImageByIdUsecase {
+class BuyMeDeleteImageByIdUsecase {
   final FolderDetailRepository repository;
 
-  DeleteImageByIdUsecase(this.repository);
+  BuyMeDeleteImageByIdUsecase(this.repository);
 
   Future<Either<APIErrors, bool>> call(String imageId) {
     return repository.deleteImageById(imageId);