+++ lib/features/folder_details/presentation/widgets/controller/folder_detail_controller.dart
@@ -3,7 +3,6 @@ import 'dart:async';
 import '../../../../../enum/app_state.dart';
 import '../../../../camera/data/models/damage_assessment_response.dart';
 import '../../../../common/base_controller.dart';
-import 'package:camera/camera.dart';
 import 'package:get/get.dart';
 
 import '../../../data/models/buy_me_image_model.dart';
@@ -11,11 +10,9 @@ import '../../../data/models/check_car_model.dart';
 import '../../../domain/usecase/check_is_one_car_usecase.dart';
 import '../../../domain/usecase/get_image_info_usecase.dart';
 
-List<CameraDescription> cameras = <CameraDescription>[];
-
-class FolderDetailController extends BaseController {
-  final GetImageInfoUsecase getImageInfoUsecase = Get.find();
-  final CheckIsOneCarUsecase checkIsOneCarUsecase = Get.find();
+class BuyMeFolderDetailController extends BuyMeBaseController {
+  final BuyMeGetImageInfoUsecase getImageInfoUsecase = Get.find();
+  final BuyMeCheckIsOneCarUsecase checkIsOneCarUsecase = Get.find();
   var checkCarModel = Rx<CheckCarModel?>(null);
   var imageInfo = Rx<BuyMeImageResponse?>(null);
   // var damageResponseListener = Rx<DamageAssessmentResponse?>(null);